pyproject.toml
[tool.poetry]
name = "exteract"
version = "0.1.0"
description = ""
authors = ["Ari <me@ari.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
ctransformers = "^0.2.27"
faiss-cpu = "^1.7.4"
langchain = "^0.0.317"
pypdf = "^3.16.4"
python-box = "^7.1.1"
sentence-transformers = "^2.2.2"
pdfminer = "^20191125"
pyyaml = "^6.0.1"
unstructured = "^0.10.24"
pdf2image = "^1.16.3"
pytesseract = "^0.3.10"
opencv-python = "^4.8.1.78"
matplotlib = "^3.8.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.25.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
